This commit was manufactured by cvs2svn to create branch 'gfs_xio_data'.
Sprout from gfork3 2007-06-05 22:14:45 UTC cvs2git 'This commit was manufactured by cvs2svn to create branch 'gfork3'.'
Cherrypick from master 2007-10-03 18:08:24 UTC bresnaha 'added a server side configuration option to disable commands.':
    common/source/dirt.sh
    common/source/library/globus_callback_threads.c
    common/source/library/globus_common_include.h
    common/source/library/globus_error_generic.c
    common/source/library/globus_hashtable.c
    common/source/library/globus_logging.c
    common/source/library/globus_logging.h
    common/source/library/globus_options.c
    common/source/library/globus_options.h
    common/source/pkgdata/pkg_data_src.gpt.in
    gridftp/client/source/dirt.sh
    gridftp/client/source/globus_ftp_client.h
    gridftp/client/source/globus_ftp_client_attr.c
    gridftp/client/source/globus_ftp_client_handle.c
    gridftp/client/source/globus_ftp_client_state.c
    gridftp/client/source/globus_i_ftp_client.h
    gridftp/client/source/pkgdata/pkg_data_src.gpt.in
    gridftp/client/test/Makefile.am
    gridftp/client/test/dirt.sh
    gridftp/client/test/globus-ftp-client-run-tests.pl
    gridftp/client/test/globus_ftp_client_pipelined_transfer_test.c
    gridftp/control/source/dirt.sh
    gridftp/control/source/globus_ftp_control.h
    gridftp/control/source/globus_ftp_control_data.c
    gridftp/gfork/source/README.txt
    gridftp/gfork/source/globus_gfork.h
    gridftp/gfork/source/globus_gfork_lib.c
    gridftp/gfork/source/globus_gfork_opts.c
    gridftp/gfork/source/globus_gfork_server.c
    gridftp/gfork/source/globus_gfork_states.c
    gridftp/gfork/source/globus_i_gfork.h
    gridftp/gfork/test_plugin/Makefile.am
    gridftp/gridftp_driver/source/Makefile.am
    gridftp/gridftp_driver/source/dirt.sh
    gridftp/gridftp_driver/source/pkgdata/pkg_data_src.gpt.in
    gridftp/gwtftp/source/Makefile.am
    gridftp/gwtftp/source/README.txt
    gridftp/gwtftp/source/bootstrap
    gridftp/gwtftp/source/configure.in
    gridftp/gwtftp/source/globus_i_gwtftp.c
    gridftp/gwtftp/source/globus_i_gwtftp.h
    gridftp/gwtftp/source/globus_i_gwtftp_client.c
    gridftp/gwtftp/source/globus_i_gwtftp_cmd.c
    gridftp/gwtftp/source/globus_i_gwtftp_data.c
    gridftp/gwtftp/source/globus_i_gwtftp_opts.c
    gridftp/gwtftp/source/globus_i_gwtftp_server.c
    gridftp/gwtftp/source/pkgdata/pkg_data_src.gpt.in
    gridftp/gwtftp/source/version.h.in
    gridftp/server-lib/src/dirt.sh
    gridftp/server-lib/src/globus_gridftp_server_control.c
    gridftp/server-lib/src/globus_gridftp_server_control.h
    gridftp/server-lib/src/globus_gridftp_server_control_commands.c
    gridftp/server-lib/src/globus_i_gridftp_server_control.h
    gridftp/server-lib/src/globus_xio_gssapi_ftp.c
    gridftp/server/acl/example/Makefile.am
    gridftp/server/acl/example/bootstrap
    gridftp/server/acl/example/configure.in
    gridftp/server/acl/example/dirt.sh
    gridftp/server/acl/example/globus_gfs_acl_example.c
    gridftp/server/acl/example/pkgdata/pkg_data_src.gpt.in
    gridftp/server/acl/example/version.h.in
    gridftp/server/doc/gfork.txt
    gridftp/server/src/Makefile.am
    gridftp/server/src/configure.in.in
    gridftp/server/src/dirt.sh
    gridftp/server/src/gfs_dynbe_client.c
    gridftp/server/src/gfs_gfork_master.c
    gridftp/server/src/gfs_i_gfork_plugin.h
    gridftp/server/src/globus_gridftp_server.c
    gridftp/server/src/globus_gridftp_server.h
    gridftp/server/src/globus_i_gfs_acl.c
    gridftp/server/src/globus_i_gfs_acl.h
    gridftp/server/src/globus_i_gfs_acl_test.c
    gridftp/server/src/globus_i_gfs_brain.c
    gridftp/server/src/globus_i_gfs_cas.c
    gridftp/server/src/globus_i_gfs_config.c
    gridftp/server/src/globus_i_gfs_config.h
    gridftp/server/src/globus_i_gfs_control.c
    gridftp/server/src/globus_i_gfs_data.c
    gridftp/server/src/globus_i_gfs_default_brain.c
    gridftp/server/src/globus_i_gfs_embed.c
    gridftp/server/src/globus_i_gfs_ipc.c
    gridftp/server/src/globus_i_gfs_log.c
    gridftp/server/src/globus_i_gfs_log.h
    gridftp/server/src/globus_i_gridftp_server.h
    gridftp/server/src/gridftp-be.gfork.in
    gridftp/server/src/gridftp-fe.gfork.in
    gridftp/server/src/gridftp.gfork.in
    gridftp/server/src/modules/remote/globus_gridftp_server_remote.c
    gridftp/server/src/pkgdata/pkg_data_src.gpt.in
    gridftp/server/test/Makefile.am
    gridftp/server/test/bootstrap
    gridftp/server/test/configure.in
    gridftp/server/test/gfork_be_conf.in
    gridftp/server/test/gfork_conf.in
    gridftp/server/test/gfork_mem_conf.in
    gridftp/server/test/gfork_single_conf.in
    gridftp/server/test/globus-gridftp-be-gfork.pl
    gridftp/server/test/globus-gridftp-dyn-be.pl
    gridftp/server/test/globus-gridftp-gfork-fe.pl
    gridftp/server/test/globus-gridftp-gfork-mem.pl
    gridftp/server/test/pkgdata/pkg_data_src.gpt.in
    gridftp/sshftp/setup/setup-globus-gridftp-sshftp.pl
    xio/drivers/blockz/Makefile.am
    xio/drivers/blockz/bootstrap
    xio/drivers/blockz/configure.in
    xio/drivers/blockz/doxygen/.cvsignore
    xio/drivers/blockz/globus_xio_blockz.c
    xio/drivers/blockz/globus_xio_blockz.h
    xio/drivers/blockz/pkgdata/pkg_data_src.gpt.in
    xio/drivers/blockz/version.h.in
    xio/drivers/gsi/globus_xio_gsi.c
    xio/drivers/netlog/Makefile.am
    xio/drivers/netlog/globus_xio_netlogger.c
    xio/drivers/udt_ref/globus_xio_udt_ref.cpp
    xio/drivers/udt_ref/globus_xio_udt_ref.h
    xio/src/builtins/http/globus_xio_http_server.c
    xio/src/builtins/tcp/globus_xio_tcp_driver.c
    xio/src/builtins/tcp/globus_xio_tcp_driver.h
    xio/src/builtins/telnet/globus_xio_telnet.c
    xio/src/dirt.sh
    xio/src/globus_xio_server.c
    xio/src/pkgdata/pkg_data_src.gpt.in
    xio/test/dirt.sh
    xio/test/http_get_test.c
    xio/test/http_post_test.c
    xio/test/http_put_test.c
    xio/test/http_test_common.c
    xio/test/http_timeout_test.c
    xio/test/pkgdata/pkg_data_src.gpt.in
Delete:
    gridftp/gwtftp/source/dirt.sh
    gridftp/server/test/Makefile

+++ gridftp/gwtftp/source/globus_i_gwtftp.c
@@ -1,4 +1,11 @@
 #include "globus_i_gwtftp.h"
+#include <sys/types.h>
+#include <sys/wait.h>
+
+#define GWTFTP_CHILD_CLIENT_FD  "GWTFTP_CHILD_CLIENT_FD"
+#define GWTFTP_CHILD_SERVER_URL "GWTFTP_CHILD_SERVER_URL"
+
+#define URL_SUBJECT_DELIM       "##"
 
 /* 
  *  This file contains the main routine, the arugment parsing and 
@@ -6,6 +13,8 @@
  *  it hands the new connection to connect.c for the FTP handlshake processing
  */
 
+char *                                  gwtftp_i_data_interface = NULL;
+globus_xio_driver_t                     gwtftp_l_gssapi_driver;
 globus_xio_driver_t                     gwtftp_l_tcp_driver;
 globus_xio_driver_t                     gwtftp_l_gsi_driver;
 globus_xio_stack_t                      gwtftp_l_data_tcp_stack;
@@ -15,7 +24,6 @@ static globus_xio_stack_t               gwtftp_l_client_stack;
 static globus_xio_stack_t               gwtftp_l_server_stack;
 static globus_xio_stack_t               gwtftp_l_client_stack;
 static globus_xio_driver_t              gwtftp_l_telnet_driver;
-static globus_xio_driver_t              gwtftp_l_gssapi_driver;
 static globus_xio_server_t              gwtftp_l_server;
 
 static globus_mutex_t                   gwtftp_l_mutex;
@@ -26,10 +34,12 @@ static globus_bool_t                    gwtftp_l_child = GLOBUS_FALSE;
 static int                              gwtftp_l_log_level = 255;
 static FILE *                           gwtftp_l_log_fptr;
 static const char *                     gwtftp_l_pw_file = NULL;
+static const char *                     gwtftp_l_log_filename = NULL;
 static int                              gwtftp_l_listen_port = 0;
 static globus_list_t *                  gwtftp_l_ip_list;
 
-static char **                          gwtftp_l_exec_program;
+static globus_bool_t                    gwtftp_l_allow_all = GLOBUS_FALSE;
+static char  *                          gwtftp_l_exec_program;
 
 static globus_list_t *                  gwtftp_l_connection_list = NULL;
 
@@ -116,10 +126,26 @@ gwtftp_l_accept_cb(
             goto error_over;
         }
 
-        /* only pass it on to the next level after we know we are not
-            immediately about to kill it off */
-        gwtftp_i_new_connection(handle);
+        /* verify that we allow this IP  */
+        result = gwtftp_i_ip_ok(handle);
+        if(result != GLOBUS_SUCCESS)
+        {
+            globus_object_t *           obj;
+            char *                      err_s;
 
+            obj = globus_error_get(result);
+            err_s = globus_error_print_friendly(obj);
+            globus_object_free(obj);
+            gwtftp_i_log(FTP2GRID_LOG_ERROR,
+                "Rejecting a client based in IP: %s\n", err_s);
+            globus_free(err_s);
+        }
+        else
+        {
+            /* only pass it on to the next level after we know we are not
+                immediately about to kill it off */
+            gwtftp_i_new_connection(handle, NULL);
+        }
     }
     globus_mutex_unlock(&gwtftp_l_mutex);
 
@@ -265,6 +291,117 @@ gwtftp_l_interrupt_cb(
 
 static
 globus_result_t
+gwtftp_l_child_start()
+{
+    char *                              tmp_env;
+    int                                 fd;
+    int                                 sc;
+    globus_xio_attr_t                   xio_attr;
+    globus_xio_handle_t                 handle;
+    globus_xio_handle_t                 server_xio;
+    globus_result_t                     result;
+    char *                              cs;
+
+    globus_xio_attr_init(&xio_attr);
+
+    globus_mutex_lock(&gwtftp_l_mutex);
+    {
+        result = globus_xio_attr_cntl(xio_attr, gwtftp_l_telnet_driver,
+            GLOBUS_XIO_TELNET_BUFFER, GLOBUS_TRUE);
+        if(result != GLOBUS_SUCCESS)
+        {
+            goto error_telnet;
+        }
+
+        cs = globus_libc_getenv(GWTFTP_CHILD_SERVER_URL);
+        if(cs == NULL)
+        {
+            goto error_env;
+        }
+
+        /* get the FD for the client CC */
+        tmp_env = globus_libc_getenv(GWTFTP_CHILD_CLIENT_FD);
+        if(tmp_env == NULL)
+        {
+            goto error_env;
+        }
+        sc = sscanf(tmp_env, "%d", &fd);
+        if(sc != 1)
+        {
+            goto error_env;
+        }
+        result = globus_xio_attr_cntl(
+            xio_attr,
+            gwtftp_l_tcp_driver,
+            GLOBUS_XIO_TCP_SET_HANDLE,
+            fd);
+        if(result != GLOBUS_SUCCESS)
+        {
+            goto error_tcp;
+        }
+        result = globus_xio_attr_cntl(
+            xio_attr,
+            gwtftp_l_telnet_driver,
+            GLOBUS_XIO_TELNET_FORCE_SERVER,
+            GLOBUS_TRUE);
+        if(result != GLOBUS_SUCCESS)
+        {
+            goto error_fs;
+        }
+
+        result = globus_xio_handle_create(&handle, gwtftp_l_client_stack);
+        if(result != GLOBUS_SUCCESS)
+        {
+            goto error_create;
+        }
+        result = globus_xio_open(handle, NULL, xio_attr);
+        if(result != GLOBUS_SUCCESS)
+        {
+            goto error_create;
+        }
+
+        globus_list_insert(&gwtftp_l_connection_list, handle);
+
+        result = globus_xio_handle_create(
+            &server_xio, gwtftp_l_server_stack);
+        if(result != GLOBUS_SUCCESS)
+        {
+            goto error_scheme;
+        }
+
+
+        result = gwtftp_i_server_conn_open(server_xio, cs, handle, NULL);
+        if(result != GLOBUS_SUCCESS)
+        {
+            goto error_open;
+        }
+        globus_list_insert(&gwtftp_l_connection_list, server_xio);
+
+        while(!globus_list_empty(gwtftp_l_connection_list))
+        {
+            globus_cond_wait(&gwtftp_l_cond, &gwtftp_l_mutex);
+        }
+        globus_xio_attr_destroy(xio_attr);
+    }
+    globus_mutex_unlock(&gwtftp_l_mutex);
+
+    return GLOBUS_SUCCESS;
+
+error_open:
+error_scheme:
+error_create:
+error_fs:
+error_tcp:
+error_env:
+error_telnet:
+    globus_mutex_unlock(&gwtftp_l_mutex);
+    globus_xio_attr_destroy(xio_attr);
+
+    return result;
+}
+
+static
+globus_result_t
 gwtftp_l_master_start()
 {
     globus_xio_attr_t                   xio_attr;
@@ -387,6 +524,8 @@ gwtftp_i_close(
             globus_list_remove(&gwtftp_l_connection_list, list);
 
             globus_xio_register_close(handle, NULL, close_cb, user_arg);
+
+            globus_cond_signal(&gwtftp_l_cond);
         }
     }
     globus_mutex_unlock(&gwtftp_l_mutex);
@@ -398,7 +537,6 @@ gwtftp_i_options(
     int                                 argc,
     char **                             argv)
 {
-    int                                 i;
     globus_i_gwtftp_cmd_opts_t          cmd_opts;
     globus_options_handle_t             opt_h;
     globus_result_t                     result;
@@ -423,8 +561,7 @@ gwtftp_i_options(
     }
     if(cmd_opts.log_file != NULL)
     {
-        gwtftp_l_log_fptr = fopen(cmd_opts.log_file, "w");
-
+        gwtftp_l_log_fptr = fopen(cmd_opts.log_file, "a");
         if(gwtftp_l_log_fptr == NULL)
         {
             gwtftp_l_log_fptr = stderr;
@@ -436,8 +573,9 @@ gwtftp_i_options(
         }
     }
 
+    gwtftp_l_child = cmd_opts.child;
     gwtftp_l_pw_file = cmd_opts.pw_file;
-    if(gwtftp_l_pw_file == NULL)
+    if(gwtftp_l_pw_file == NULL && !gwtftp_l_child)
     {
         /* WARN */
         gwtftp_i_log(FTP2GRID_LOG_MUST,
@@ -455,22 +593,10 @@ gwtftp_i_options(
         globus_list_insert(&gwtftp_l_ip_list, (void *)"[::1]");
     }
 
-    gwtftp_l_exec_program = (char **) globus_calloc(argc+2, sizeof(char *));
-    if(gwtftp_l_exec_program == NULL)
-    {
-        result = GlobusFTP2GridError("Small malloc failure.",
-            GLOBUS_FTP2GRID_ERROR_MALLOC);
-        goto error;
-    }
-    for(i = 0; i < argc; i++)
-    {
-        gwtftp_l_exec_program[i] = argv[i];
-    }
-    gwtftp_l_exec_program[i] = "-CH";
-    i++;
-    gwtftp_l_exec_program[i] = NULL;
+    gwtftp_l_log_filename = cmd_opts.log_file;
+    gwtftp_l_exec_program = argv[0];
 
-    gwtftp_l_child = cmd_opts.child;
+    gwtftp_l_allow_all = cmd_opts.allow_all;
 
     globus_options_destroy(opt_h);
 
@@ -480,13 +606,59 @@ error:
     return result;
 }
 
+static
+void
+gwtftp_l_sigchld(
+    void *                              user_arg)
+{
+    int                                 child_pid;
+    int                                 child_rc;
+    int                                 child_status;
+
+    while((child_pid = waitpid(-1, &child_status, WNOHANG)) > 0)
+    {
+        if(WIFEXITED(child_status))
+        {
+            /* normal exit */
+            child_rc = WEXITSTATUS(child_status);
+        }
+        if(child_rc != 0)
+        {
+            gwtftp_i_log(
+                FTP2GRID_LOG_WARN, "Child pid %d finished with rc %d\n",
+                child_pid, child_rc);
+        }
+    }
+#   ifdef BUILD_LITE
+    {
+        globus_callback_register_signal_handler(
+            SIGCHLD,
+            GLOBUS_FALSE,
+            gwtftp_l_sigchld,
+            user_arg);
+    }
+#   else
+    {
+        globus_reltime_t            delay;
+
+        GlobusTimeReltimeSet(delay, 5, 0);
+
+        globus_callback_register_oneshot(
+            NULL,
+            &gfork_l_sigchild_fake,
+            gwtftp_l_sigchld,
+            user_arg);
+    }
+#   endif
+}
+
 int
 main(
     int                                 argc,
     char **                             argv)
 {
     globus_result_t                     result;
-
+    
     globus_module_activate(GLOBUS_XIO_MODULE);
     globus_module_activate(GLOBUS_GRIDFTP_SERVER_CONTROL_MODULE);
 
@@ -507,9 +679,36 @@ main(
     gwtftp_i_server_init();
     if(gwtftp_l_child)
     {
+        /* convert FDs to a handle */
+        result = gwtftp_l_child_start();
+        if(result != GLOBUS_SUCCESS)
+        {
+            goto error;
+        }
     }
     else
     {
+        if(gwtftp_l_daemon)
+        {
+#           ifdef BUILD_LITE
+            {
+                result = globus_callback_register_signal_handler(
+                    SIGCHLD,
+                    GLOBUS_FALSE,
+                    gwtftp_l_sigchld,
+                    NULL);
+            }
+#           else
+            {
+                result = globus_callback_register_oneshot(
+                    NULL,
+                    &gfork_l_sigchild_fake,
+                    gwtftp_l_sigchld,
+                    NULL);
+            }
+#       endif
+        }
+
         result = gwtftp_l_master_start();
         if(result != GLOBUS_SUCCESS)
         {
@@ -589,6 +788,11 @@ gwtftp_i_ip_ok(
     globus_bool_t                       ok = GLOBUS_FALSE;
     globus_list_t *                     list;
     GlobusFTP2GridFuncName(gwtftp_i_ip_ok);
+
+    if(gwtftp_l_allow_all)
+    {
+        return GLOBUS_SUCCESS;
+    }
 
     result = globus_xio_handle_cntl(
         handle,
@@ -672,21 +876,24 @@ gwtftp_i_exec__unix(
     globus_xio_handle_t                 client_xio,
     globus_url_t *                      url)
 {
+    int                                 i;
     globus_xio_system_socket_t          socket_handle;
     pid_t                               pid;
     int                                 rc;
+    char                                env_str[64];
+    char *                              exec_program[32];
+    char *                              cs;
     globus_result_t                     result;
 
+    gwtftp_i_log(FTP2GRID_LOG_INFO, "Forking a child.\n");
     pid = fork();
     if(pid == 0)
     {
-        /* do i set the uid or do i let a sudo thing do it */
-        rc = setuid(uid);
-        if(rc != 0)
-        {
-            gwtftp_i_log(FTP2GRID_LOG_WARN, "Child setuid failed.\n");
-            exit(rc);
-        }
+        cs = globus_common_create_string("%s:%d", url->host, url->port);
+        globus_libc_setenv(GWTFTP_CHILD_SERVER_URL, cs, 1);
+
+        gwtftp_i_log(FTP2GRID_LOG_INFO, "Child processing seting uid to %d.\n",
+            uid);
         /* extract the FD */
         result = globus_xio_handle_cntl(
             client_xio,
@@ -698,20 +905,52 @@ gwtftp_i_exec__unix(
             gwtftp_i_log(FTP2GRID_LOG_WARN, "Failed to extract socket.\n");
             exit(1);
         }
-        rc = dup2(socket_handle, STDIN_FILENO);
-        if(rc < 0)
+        /* all xio sockets will close on exec */
+        sprintf(env_str, "%d", socket_handle);
+        globus_libc_setenv(GWTFTP_CHILD_CLIENT_FD, env_str, 1);
+
+        exec_program[0] = strdup(gwtftp_l_exec_program);
+        exec_program[1] = strdup("-CH");
+        i = 2;
+        if(gwtftp_l_log_filename != NULL)
         {
-            gwtftp_i_log(FTP2GRID_LOG_WARN, "Failed to dup2 socket.\n");
+            exec_program[i] = strdup("-l");
+            i++;
+            exec_program[i] = globus_common_create_string("%s.%d", 
+                gwtftp_l_log_filename, getpid());
+            i++;
+        }
+        exec_program[i] = strdup("-lm");
+        i++;
+        exec_program[i] = globus_common_create_string("%d", gwtftp_l_log_level);
+        i++;
+        exec_program[i] = NULL;
+
+        gwtftp_i_log(FTP2GRID_LOG_INFO, "Execing child process: ");
+        for(i = 0; exec_program[i] != NULL; i++)
+        {
+            gwtftp_i_log(FTP2GRID_LOG_INFO, "%s ", exec_program[i]);
+        }
+        gwtftp_i_log(FTP2GRID_LOG_INFO, "\n");
+        rc = setuid(uid);
+        if(rc != 0)
+        {
+            gwtftp_i_log(FTP2GRID_LOG_WARN, "Child setuid failed.\n");
             exit(rc);
         }
-        /* all xio sockets will close on exec */
+        fflush(gwtftp_l_log_fptr);
 
-        rc = execvp(gwtftp_l_exec_program[0], gwtftp_l_exec_program);
+        rc = execvp(exec_program[0], exec_program);
+        for(i = 0; exec_program[i] != NULL; i++)
+        {
+            free(exec_program[i]);
+        }
+        gwtftp_i_log(FTP2GRID_LOG_ERROR, "Exec failed\n");
         exit(rc);
     }
     else if(pid < 0)
     {
-        gwtftp_i_log(FTP2GRID_LOG_INFO, "Failed to fork\n");
+        gwtftp_i_log(FTP2GRID_LOG_ERROR, "Failed to fork\n");
     }
 }
 
@@ -795,6 +1034,7 @@ gwtftp_i_authorized_user(
     globus_url_t                        g_url;
     char *                              url_retry = NULL;
     char *                              tmp_ptr;
+    char *                              subject = NULL;
 
     gwtftp_i_log(FTP2GRID_LOG_INFO,
         "Authorizing: %s\n", full_username);
@@ -802,6 +1042,18 @@ gwtftp_i_authorized_user(
     {
         list = globus_list_search(gwtftp_l_connection_list, client_xio);
         globus_assert(list != NULL);
+
+        /* pull out local user name and  subject if they are there */
+
+        tmp_ptr = strstr(full_username, URL_SUBJECT_DELIM);
+        if(tmp_ptr != NULL)
+        {
+            subject = strdup(full_username);
+            full_username = tmp_ptr + strlen(URL_SUBJECT_DELIM);
+            tmp_ptr = strstr(subject, URL_SUBJECT_DELIM);
+            globus_assert(tmp_ptr != NULL);
+            *tmp_ptr = '\0';
+        }
 
         rc = globus_url_parse(full_username, &g_url);
         if(rc == GLOBUS_URL_ERROR_BAD_SCHEME)
@@ -868,7 +1120,7 @@ gwtftp_i_authorized_user(
                     "%s:%d", g_url.host, g_url.port);
 
                 result = gwtftp_i_server_conn_open(
-                    server_xio, cs, client_xio);
+                    server_xio, cs, client_xio, subject);
                 globus_free(cs);
                 if(result != GLOBUS_SUCCESS)
                 {
